move calc superobservation error to utils
