+++ src/inversion_scripts/utils.py
@@ -342,3 +342,24 @@ def calculate_area_in_km(coordinate_list):
     poly_area, _ = geod.geometry_area_perimeter(polygon)
 
     return abs(poly_area) * 1e-6
+
+def calculate_superobservation_error(sO, p):
+    """
+    Returns the estimated observational error accounting for superobservations.
+    Using eqn (5) from Chen et al., 2023, https://doi.org/10.5194/egusphere-2022-1504
+    Args:
+        sO : float
+            observational error specified in config file
+        p  : float
+            average number of observations contained within each superobservation
+    Returns:
+         s_super: float
+            observational error for superobservations
+    """
+    # values from Chen et al., 2023, https://doi.org/10.5194/egusphere-2022-1504
+    r_retrieval = 0.55
+    s_transport = 4.5
+    s_super = np.sqrt(
+        sO**2 * (((1 - r_retrieval) / p) + r_retrieval) + s_transport**2
+    )
+    return s_super
